fix: change hero and appbar names completely to try to fix this remote subdir casingbug
