--- src/components/templates/Hero/index.ts
@@ -1,1 +1,0 @@
-export * from "./Hero";
