--- src/components/templates/appBar/AppBar.module.css
@@ -1,4 +1,0 @@
-.appbar {
-    background-color: transparent !important;
-    box-shadow: none !important;
-}
