--- src/components/templates/appBar/index.ts
@@ -1,1 +1,0 @@
-export * from "./AppBar";
