--- src/components/templates/hero/Hero.module.css
@@ -1,38 +1,0 @@
-.container {
-    background-size: cover;
-    width: 100%;
-    height: 100vh;
-    background-image: linear-gradient(
-            rgba(8, 122, 217, 0.8),
-            rgba(8, 122, 217, 0.6)
-        ),
-        url(/hero.png);
-    background-position: center no-repeat;
-
-    overflow-y: hidden;
-    margin: auto;
-    align-items: center;
-    display: flex;
-}
-
-.full-screen-text {
-    display: block !important;
-}
-
-.mobile-screen-text {
-    display: none !important;
-}
-
-@media (max-width: 815px) {
-    .centered-text {
-        width: 70vw;
-    }
-
-    .full-screen-text {
-        display: none !important;
-    }
-
-    .mobile-screen-text {
-        display: block !important;
-    }
-}
