+++ src/components/templates/landing/index.ts
@@ -1,0 +1,1 @@
+export * from "./Landing";
