+++ src/components/templates/navBar/index.ts
@@ -1,0 +1,1 @@
+export * from "./NavBar";
