+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import { SectionContainer } from "@/components/organisms/sectionContainer";
-import { Hero } from "@/components/templates/hero";
+import { Landing } from "@/components/templates/landing";
 import { SnapSectionContainer } from "@/components/templates/snapSectionContainer";
 import "animate.css";
 import Head from "next/head";
@@ -10,7 +10,7 @@ export default function Index() {
       <Head>
         <title>Ben Horner&apos;s Portfolio</title>
       </Head>
-      <Hero />
+      <Landing />
       <SnapSectionContainer>
         <SectionContainer />
       </SnapSectionContainer>
